+++ src/renderer/api/obs/providers/isStreaming.ts
@@ -18,6 +18,7 @@ export interface StreamingState {
 // Provider
 // ============================================================================
 
+// Export here and in index.ts for dynamic loading by name
 export const isStreaming = (obs: ConnectionPublic) => createProvider({
 	init: (onChanged) => {
 		const setData = (state: string) => {