+++ src/renderer/api/obs/providers/stats.ts
@@ -6,6 +6,7 @@ import { Stats } from '../abstraction/types'
 // Provider
 // ============================================================================
 
+// Export here and in index.ts for dynamic loading by name
 export const stats = (obs: ConnectionPublic, {
 	refreshTime = 3000,
 } = {}) => createProvider({